refactor(frontend): no more topbar in layout per feature but each page decide if they want to place their own header
