+++ apps/frontend/src/modules/core/presentation/models/RouteHandle.ts
@@ -1,12 +1,8 @@
-import type { ReactNode } from 'react';
-
 export interface BreadcrumbParams {
   pipelineId?: string;
   projectId?: string;
 }
 
 export interface RouteHandle {
-  topbar?: ReactNode;
-  tabsDefaultValue?: string;
   breadcrumb?: string | ((params: BreadcrumbParams) => string);
 }
