+++ apps/frontend/src/modules/core/presentation/ui/CoreRouter.tsx
@@ -4,14 +4,10 @@ import LoginPage from '@/modules/features/login/presentation/ui/LoginPage';
 import UserSettingsPage from '@/modules/features/user_settings/presentation/ui/UserSettingsPage';
 import MarketplacePage from '@/modules/features/marketplace/presentation/ui/MarketplacePage';
 import { Layout } from '@/modules/layout/presentation/ui/Layout.tsx';
-import { MarketplaceTopBar } from '@/modules/features/marketplace/presentation/ui/MarketplaceTopBar';
 import { RequireAuth } from '@core/presentation/ui/RequireAuth.tsx';
 import { PipelineCreationPage } from '@/modules/features/pipeline-creation/presentation/ui/PipelineCreationPage.tsx';
-import { PipelineCreationTopbar } from '@/modules/features/pipeline-creation/presentation/ui/PipelineCreationTopbar.tsx';
 import { DashboardPipelinePage } from '@/modules/features/pipeline-dashboard/presentation/ui/DashboardPipelinePage';
-import { PipelineDashboardTopBar } from '@/modules/features/pipeline-dashboard/presentation/ui/PipelineDashboardTopBar.tsx';
 import ProjectPage from '@/modules/features/project/presentation/ui/ProjectPage.tsx';
-import { ProjectTopBar } from '@/modules/features/project/presentation/ui/ProjectTopBar.tsx';
 import type { BreadcrumbParams } from '@core/presentation/models/RouteHandle.ts';
 
 //TODO: put each navigations part in a separate file, (module ?)
@@ -32,48 +28,46 @@ export const CoreRouter = createBrowserRouter([
             element: <UserSettingsPage />,
           },
           {
-            path: '/projects/:projectId/',
-            handle: { breadcrumb: (params: BreadcrumbParams) => `Project #${params.projectId}` },
+            path: '/projects',
+            handle: {
+              breadcrumb: () => 'Projects',
+            },
             children: [
               {
                 index: true,
-                element: <DashboardPipelinePage />,
-                handle: {
-                  topbar: <PipelineDashboardTopBar />,
-                },
+                element: <ProjectPage />,
               },
               {
-                path: 'create',
-                element: <PipelineCreationPage />,
+                path: ':projectId',
                 handle: {
-                  breadcrumb: (_: BreadcrumbParams) => `Create`,
-                  topbar: <PipelineCreationTopbar />,
-                  tabsDefaultValue: 'scripting',
+                  breadcrumb: (params: BreadcrumbParams) => `Project #${params.projectId}`,
                 },
-              },
-              {
-                path: 'edit/:pipelineId',
-                element: <PipelineCreationPage />,
-                handle: {
-                  topbar: <PipelineCreationTopbar />,
-                  tabsDefaultValue: 'scripting',
-                },
+                children: [
+                  {
+                    index: true,
+                    element: <DashboardPipelinePage />,
+                  },
+                  {
+                    path: 'create',
+                    element: <PipelineCreationPage />,
+                    handle: {
+                      breadcrumb: () => `Create`,
+                    },
+                  },
+                  {
+                    path: 'edit/:pipelineId',
+                    element: <PipelineCreationPage />,
+                    handle: {
+                      breadcrumb: ({ pipelineId }: BreadcrumbParams) => `Pipeline #${pipelineId}`,
+                    },
+                  },
+                ],
               },
             ],
           },
           {
             path: '/marketplace',
             element: <MarketplacePage />,
-            handle: {
-              topbar: <MarketplaceTopBar />,
-            },
-          },
-          {
-            path: '/projects',
-            element: <ProjectPage />,
-            handle: {
-              topbar: <ProjectTopBar />,
-            },
           },
         ],
       },
