+++ apps/frontend/src/modules/features/project/presentation/ui/ProjectHeader.tsx
@@ -1,0 +1,30 @@
+import { Button } from '@shadcn';
+import { AddProjectDialog } from '@/modules/features/project/presentation/ui/AddProjectDialog.tsx';
+import { useState } from 'react';
+
+interface ProjectHeaderProps {
+  numberOfProjects: number;
+}
+
+export const ProjectHeader = ({ numberOfProjects }: ProjectHeaderProps) => {
+  const [open, setOpen] = useState(false);
+
+  return (
+    <div className={'flex flex-row justify-between items-center w-full'}>
+      <div className='flex items-baseline gap-2'>
+        <h1 className='text-3xl font-bold tracking-tight'>
+          <span className='text-primary'>{numberOfProjects}</span>{' '}
+          <span className='text-foreground'>
+            Project
+            {numberOfProjects > 1 ? 's' : ''}
+          </span>
+        </h1>
+        <span className='text-sm text-muted-foreground font-medium'>in total</span>
+      </div>
+      <Button onClick={() => setOpen(true)} variant={'default'} className={'ml-auto'}>
+        New project
+      </Button>
+      <AddProjectDialog open={open} setOpen={setOpen} />
+    </div>
+  );
+};
