+++ apps/frontend/src/modules/features/project/presentation/ui/ProjectPage.tsx
@@ -3,6 +3,7 @@ import ProjectCard from '@/modules/features/project/presentation/ui/ProjectCard.
 import ProjectCardSkeleton from '@/modules/features/project/presentation/ui/ProjectCardSkeleton.tsx';
 import { Skeleton } from '@/modules/shared/presentation/ui/shadcn/skeleton.tsx';
 import { useDelayedLoading } from '@/modules/shared/presentation/hooks/useDelayedLoading.ts';
+import { ProjectHeader } from '@/modules/features/project/presentation/ui/ProjectHeader.tsx';
 
 export const ProjectPage = () => {
   const { projects, isLoading, isError } = useProjects();
@@ -41,16 +42,7 @@ export const ProjectPage = () => {
 
   return (
     <div className='flex flex-col gap-4 w-full h-full p-2'>
-      <div className='flex items-baseline gap-2'>
-        <h1 className='text-3xl font-bold tracking-tight'>
-          <span className='text-primary'>{projects.length}</span>{' '}
-          <span className='text-foreground'>
-            Project
-            {projects.length > 1 ? 's' : ''}
-          </span>
-        </h1>
-        <span className='text-sm text-muted-foreground font-medium'>in total</span>
-      </div>
+      <ProjectHeader numberOfProjects={projects.length} />
       <div className='grid grid-cols-1 sm:grid-cols-2 md:grid-cols-3 lg:grid-cols-4 gap-4'>
         {projects &&
           projects.map(project => <ProjectCard key={project.projectId} project={project} />)}
